Independent Widget class for Your pronouns and Im looking for cards for reusability
