+++ lib/design/components/onboarding/exports.dart
@@ -1,5 +1,7 @@
 export 'glint_custom_icon_chip.dart';
+export 'im_looking_for_card.dart';
 export 'interest_and_vibe_selection.dart';
 export 'onboarding_safe_area_container.dart';
 export 'upload_picture_container.dart';
 export 'upload_pictures_containers.dart';
+export 'your_pronouns_card.dart';
